moved functions to new below_ground file
