+++ docs/source/api/hydrology/below_ground.md
@@ -1,0 +1,24 @@
+---
+jupytext:
+  cell_metadata_filter: -all
+  formats: md:myst
+  main_language: python
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.13.8
+kernelspec:
+  display_name: vr_python3
+  language: python
+  name: vr_python3
+---
+
+#  API for the {mod}`~virtual_rainforest.models.hydrology.below_ground` module
+
+```{eval-rst}
+.. automodule:: virtual_rainforest.models.hydrology.below_ground
+    :autosummary:
+    :members:
+    :special-members: __init__
+```
